fix: address review findings on pr #30886
